test(core): split edge combination tests
